refactor(sight): split genai/builder.rs into 4 focused modules (#950)
Split the 2225-line builder.rs (exceeding AGENTS.md 2000-line threshold) into
4 single-responsibility modules:

- builder.rs (502 lines): orchestration entry point
- call_builder.rs (609 lines): LLM call construction
- openai_parse.rs (685 lines): OpenAI format parsing & SSE aggregation
- helpers.rs (491 lines): route matching, provider/model extraction, message utils

Implementation details:
- Pure structural move, zero logic changes
- Tests follow their respective functions
- Uses Rust split-impl pattern (impl GenAIBuilder across files)
- Internal modules use mod (not pub mod) with pub(super) functions
- External API unchanged: pub use builder::GenAIBuilder in mod.rs

Verified: cargo build + cargo test (573 passed) + arch boundary check PASS
