+++ src/agentsight/src/genai/mod.rs
@@ -5,11 +5,14 @@
 
 pub mod anolisa_release;
 pub mod builder;
+mod call_builder;
 pub mod encrypt;
 pub mod exporter;
+mod helpers;
 pub mod id_resolver;
 pub mod instance_id;
 pub mod logtail;
+mod openai_parse;
 pub mod semantic;
 pub mod storage;
